[TimeSurfaceGenerator] split in 2 files hotsblocs.hpp timeSurfaceGenerator.hpp: able to build 1D and 2D timeSurface using events hotsBlocs.hpp: clustering classes able to cluster timeSurface
